Separate hook for client preloading
